+++ client/presentation/src/main/java/com/andone/memorip/presentation/BentoCard.kt
@@ -1,0 +1,41 @@
+package com.andone.memorip.presentation
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+
+@Composable
+fun BentoCard(item: BentoItem) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .height(item.height)
+    ) {
+        Box(
+            modifier = Modifier.fillMaxSize()
+                .padding(2.dp)
+                .clip(RoundedCornerShape(16.dp))
+                .background(Color.LightGray),
+            contentAlignment = Alignment.Center
+        ) {
+            AsyncImage(
+                model = item.imageUrl,
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+    }
+}
