+++ client/presentation/src/main/java/com/andone/memorip/presentation/BentoItem.kt
@@ -1,0 +1,10 @@
+package com.andone.memorip.presentation
+
+import androidx.compose.ui.unit.Dp
+
+data class BentoItem(
+    val id: Int,
+    val span: Int,
+    val height: Dp,
+    val imageUrl: String = ""
+)
